train/reinforcement_learning/ppo_vs_code/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tensorboard>=2.13.0
matplotlib>=3.7.0
tqdm>=4.65.0
gym>=0.26.0
